Added backend validation for books

+++ backend/src/main/java/recruitment/task/books/service/BookServiceimpl.java
@@ -10,8 +10,8 @@ import recruitment.task.books.dto.response.BookResponse;
 import recruitment.task.books.entity.Book;
 import recruitment.task.books.mapper.BookMapper;
 import recruitment.task.books.repository.BookRepository;
+import recruitment.task.books.validator.BookValidator;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Function;
@@ -20,15 +20,21 @@ import java.util.function.Function;
 public class BookServiceimpl implements BookService {
     private final BookRepository bookRepository;
     private final BookMapper bookMapper;
+    private final BookValidator bookValidator;
 
-    public BookServiceimpl(BookRepository bookRepository, BookMapper bookMapper) {
+    public BookServiceimpl(BookRepository bookRepository, BookMapper bookMapper, BookValidator bookValidator) {
         this.bookRepository = bookRepository;
         this.bookMapper = bookMapper;
+        this.bookValidator = bookValidator;
     }
 
     @Override
     public Page<BookResponse> getAllPaginated(int page, int size, Boolean order, String sortBy) {
         Pageable getPage;
+
+        if (!bookValidator.validatePagination(page, size, sortBy)) {
+            throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
+        }
 
         if (order) {
             getPage = PageRequest.of(page, size, Sort.by(Sort.Direction.ASC, sortBy));
@@ -52,6 +58,10 @@ public class BookServiceimpl implements BookService {
     public Page<BookResponse> searchBook(int page, int size, Boolean order, String sortBy, String search) {
         Pageable getPage;
         Page<Book> bookPageList;
+
+        if (!bookValidator.validatePagination(page, size, sortBy)) {
+            throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
+        }
 
         if (order) {
             getPage = PageRequest.of(page, size, Sort.by(Sort.Direction.ASC, sortBy));
@@ -87,10 +97,11 @@ public class BookServiceimpl implements BookService {
 
     @Override
     public BookResponse createBook(BookRequest bookRequest) {
-        Book book = bookMapper.mapToEntity(bookRequest);
-        if (book.getGenre() == null) {
+        if (!bookValidator.validateBook(bookRequest)) {
             throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
         }
+
+        Book book = bookMapper.mapToEntity(bookRequest);
         bookRepository.save(book);
 
         return bookMapper.entityToResponse(book);
@@ -100,6 +111,15 @@ public class BookServiceimpl implements BookService {
     public BookResponse editBook(Long id, BookRequest bookRequest) {
         Optional<Book> optionalBook = bookRepository.findById(id);
         Book book;
+
+        if (!bookValidator.validateBook(bookRequest)) {
+            throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
+        }
+
+        if (!optionalBook.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
         book = optionalBook.get();
 
         bookMapper.mapToEntity(book, bookRequest);
@@ -112,6 +132,11 @@ public class BookServiceimpl implements BookService {
     public HttpStatus deleteBook(Long id) {
         Optional<Book> optionalBook = bookRepository.findById(id);
         Book book;
+
+        if (!optionalBook.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
         book = optionalBook.get();
 
         bookRepository.delete(book);
